requirements.txt
google-genai
numpy
openai
pandas
plotly
streamlit
altair
blinker
cachetools
click
gitpython
pyarrow
pydeck
requests
tenacity
toml
tornado
watchdog